refactor: minor cleanup, error msg improvement
