+++ apps/web/i18n/locales/de/common.json
@@ -270,7 +270,6 @@
   "catchment_area_shape_tooltip": "Wählen Sie die Form der Einzugsgebietsgeometrie aus.",
   "catchment_area_type": "Art des Einzugsgebiets",
   "catchment_area_v2": "Einzugsgebiet",
-  "catchment_area_v2_description": "Einzugsgebiete zeigen, wie weit Personen innerhalb einer festgelegten Reisezeit oder Entfernung von einem oder mehreren ausgewählten Punkten reisen können.",
   "categories_chart": "Kategorien",
   "category": "Kategorie",
   "category_aggregate": "Aggregieren",
@@ -914,25 +913,23 @@
   "halo_width": "Halo-Breite",
   "heading_typography": "Überschriften",
   "heads_and_tails": "Heads und Tails",
+  "heatmap_2sfca_v2": "Heatmap 2SFCA V2",
   "heatmap_closest_average": "Heatmap Durchschnitt Reisezeit",
   "heatmap_closest_average_active_mobility": "Heatmap Durchschnitt Aktive Mobilität",
   "heatmap_closest_average_car": "Heatmap Durchschnitt Auto",
   "heatmap_closest_average_description": "Es wird die durchschnittliche Reisezeit zu den nächsten N Zielen aufgezeigt ausgehend von jeder Rasterzelle.",
   "heatmap_closest_average_v2": "Heatmap Durchschnitt Reisezeit V2",
-  "heatmap_closest_average_v2_description": "Es wird die durchschnittliche Reisezeit zu den nächsten N Zielen aufgezeigt ausgehend von jeder Rasterzelle.",
   "heatmap_computation_started": "Die Heatmap-Berechnung wurde gestartet",
   "heatmap_connectivity": "Heatmap Konnektivität",
   "heatmap_connectivity_active_mobility": "Heatmap Connectivity Aktive Mobilität",
   "heatmap_connectivity_car": "Heatmap Konnektivität Auto",
   "heatmap_connectivity_description": "Die Vernetzung der Verkehrswege wird ausgedrückt durch die Berechnung von Einzugsbereichen für jede Rasterzelle, wobei die erreichbare Fläche der Einzugsbereiche die Konnektivität ausdrückt.",
   "heatmap_connectivity_v2": "Heatmap Konnektivität V2",
-  "heatmap_connectivity_v2_description": "Die Vernetzung der Verkehrswege wird ausgedrückt durch die Berechnung von Einzugsbereichen für jede Rasterzelle, wobei die erreichbare Fläche der Einzugsbereiche die Konnektivität ausdrückt.",
   "heatmap_gravity": "Heatmap Gravity",
   "heatmap_gravity_active_mobility": "Heatmap Gravity Aktive Mobilität",
   "heatmap_gravity_car": "Heatmap Gravity Auto",
   "heatmap_gravity_description": "Es wird die Erreichbarkeit ausgehend von jeder Rasterzelle zu mehreren Orten gleichzeitig berechnet. Die Reisezeiten werden hierfür mit einer Widerstandsfunktion gewichtet und aggregiert.",
   "heatmap_gravity_v2": "Heatmap Gravity V2",
-  "heatmap_gravity_v2_description": "Es wird die Erreichbarkeit ausgehend von jeder Rasterzelle zu mehreren Orten gleichzeitig berechnet. Die Reisezeiten werden hierfür mit einer Widerstandsfunktion gewichtet und aggregiert.",
   "height": "Höhe",
   "help": "Hilfe",
   "hexagon_bin": "Hexagon Raster (H3)",
@@ -953,6 +950,7 @@
   "hover_for_help": "Mit Maus über Element fahren für Details",
   "hover_function_for_help": "Mit Maus über Funktion fahren für Details",
   "html_mode_help": "Rohes HTML schreiben. Feldwerte mit {{field_name}} einfügen und Liquid für Logik nutzen — {% if %}-Bedingungen, {% for %}-Schleifen und Filter wie {{ field | upcase }}.",
+  "huff_model_v2": "Huff-Modell V2",
   "icon_color": "Symbolfarbe",
   "icon_upload_failed": "Symbol konnte nicht hochgeladen werden",
   "icons": "Symbole",
@@ -1314,6 +1312,7 @@
   "no_layers_added": "Keine Layer hinzugefügt",
   "no_layers_found": "Keine Layer gefunden",
   "no_layouts_yet": "Noch keine Layouts",
+  "no_line_layer_found": "Kein Linien-Layer gefunden",
   "no_logo_uploaded": "Kein Logo hochgeladen",
   "no_matching_categories": "Keine passenden Kategorien",
   "no_matching_runs": "Keine passenden Ausführungen",
