+++ apps/web/i18n/locales/en/common.json
@@ -253,7 +253,6 @@
   "catalogue": "Catalogue",
   "catchment_area": "Catchment Area (Legacy)",
   "catchment_area_v2": "Catchment Area",
-  "catchment_area_v2_description": "Catchment Areas show how far people can travel within a set travel time or distance from one or more selected points.",
   "catchment_area_active_mobility": "Catchment Area Active Mobility",
   "catchment_area_car": "Catchment Area Car",
   "catchment_area_computation_started": "Catchment area computation started",
@@ -857,25 +856,23 @@
   "header_color": "Header Color",
   "heading_typography": "Headings",
   "heads_and_tails": "Heads and Tails",
+  "heatmap_2sfca_v2": "Heatmap 2SFCA V2",
   "heatmap_closest_average": "Heatmap Closest Average",
   "heatmap_closest_average_active_mobility": "Heatmap Closest Average Active Mobility",
   "heatmap_closest_average_car": "Heatmap Closest Average Car",
   "heatmap_closest_average_description": "The average travel time to the nearest N destinations is shown, starting from each grid cell.",
   "heatmap_closest_average_v2": "Heatmap Closest Average V2",
-  "heatmap_closest_average_v2_description": "The average travel time to the nearest N destinations is shown, starting from each grid cell.",
   "heatmap_computation_started": "Heatmap computation started",
   "heatmap_connectivity": "Heatmap Connectivity",
   "heatmap_connectivity_active_mobility": "Heatmap Connectivity Active Mobility",
   "heatmap_connectivity_car": "Heatmap Connectivity Car",
   "heatmap_connectivity_description": "Connectivity of transport routes is determined by calculating catchment areas for each grid cell, where the reached area is the measure for connectivity.",
   "heatmap_connectivity_v2": "Heatmap Connectivity V2",
-  "heatmap_connectivity_v2_description": "Connectivity of transport routes is determined by calculating catchment areas for each grid cell, where the reached area is the measure for connectivity.",
   "heatmap_gravity": "Heatmap Gravity",
   "heatmap_gravity_active_mobility": "Heatmap Gravity Active Mobility",
   "heatmap_gravity_car": "Heatmap Gravity Car",
   "heatmap_gravity_description": "The accessibility is calculated from each grid cell to multiple locations simultaneously. The travel times are weighted and aggregated using a impedance function.",
   "heatmap_gravity_v2": "Heatmap Gravity V2",
-  "heatmap_gravity_v2_description": "The accessibility is calculated from each grid cell to multiple locations simultaneously. The travel times are weighted and aggregated using a impedance function.",
   "height": "Height",
   "help": "Help",
   "hexagon_bin": "Hexagon grid (H3)",
@@ -894,6 +891,7 @@
   "hover_for_help": "Hover over an item to see its description",
   "hover_function_for_help": "Hover over a function to see its description",
   "html_mode_help": "Write raw HTML. Insert field values with {{field_name}}, and use Liquid for logic — {% if %} conditionals, {% for %} loops, and filters like {{ field | upcase }}.",
+  "huff_model_v2": "Huff Model V2",
   "font_color": "Font Color",
   "icon_color": "Icon Color",
   "icon_upload_failed": "Failed to upload icon",
@@ -1242,6 +1240,7 @@
   "no_layers_added": "No layers added",
   "no_layers_found": "No layers found",
   "no_layouts_yet": "No layouts yet",
+  "no_line_layer_found": "No line layer found",
   "no_matching_categories": "No matching categories",
   "no_matching_runs": "No matching runs",
   "no_measurements_yet": "No measurements yet",
